WIP: restructuring ieee754 processing to support all bfloat cases normal/subnormal, nbits less and greater than 64
