+++ benchmark/performance/arithmetic/integer/performance.cpp
@@ -98,8 +98,16 @@ try {
 
 #if MANUAL_TESTING
 
-	TestShiftOperatorPerformance();
-	TestArithmeticOperatorPerformance();
+	{
+		// test a very large integer
+		integer<1024 * 128, uint32_t> a, b, c;
+		a = 1;
+		b = 1234567890;
+		c = a * b;
+		c /= a;
+		if (c == b) cout << "PASS\n";
+	}
+
 
 	cout << "done" << endl;
