+++ include/universal/native/constexpr754.hpp
@@ -1,0 +1,97 @@
+#pragma once
+// ieee754.hpp: manipulation functions for IEEE-754 native types
+//
+// Copyright (C) 2017-2021 Stillwater Supercomputing, Inc.
+//
+// This file is part of the universal numbers project, which is released under an MIT Open Source license.
+#include <sstream>
+#include <iomanip>
+#include <cmath>    // for frexpf/frexp/frexpl  float/double/long double fraction/exponent extraction
+#include <limits>
+#include <tuple>
+
+#ifdef _MSC_VER
+#pragma warning(disable : 4127) // warning C4127: conditional expression is constant
+#endif
+#include <universal/utility/color_print.hpp>
+
+namespace sw::universal {
+
+////////////////////////////////////////////////////////////////////////
+// numerical helpers
+
+inline constexpr void extractFields(float value, bool& s, uint64_t& rawExponentBits, uint64_t& rawFractionBits) {
+	// normal number
+	uint32_t bc = std::bit_cast<uint32_t>(value);
+	s = (0x8000'0000u & bc);
+	rawExponentBits = (0x7F80'0000u & bc) >> 23;
+	rawFractionBits = (0x007F'FFFFu & bc);
+}
+
+inline constexpr void extractFields(double value, bool& s, uint64_t& rawExponentBits, uint64_t& rawFractionBits) {
+	uint64_t bc = std::bit_cast<uint64_t>(value);
+	s = (0x8000'0000'0000'0000ull & bc);
+	rawExponentBits = (0x7FF0'0000'0000'0000ull & bc) >> 52;
+	rawFractionBits = (0x000F'FFFF'FFFF'FFFFull & bc);
+}
+
+////////////////// string operators
+
+
+
+/////////////////////////////////////////////////////////////////////////////////////////////////////////
+// compiler specific long double IEEE floating point
+
+/*
+	Long double is not consistently implemented across different compilers.
+	The following section organizes the implementation details of each
+	of the compilers supported.
+
+	The x86 extended precision format is an 80-bit format first
+	implemented in the Intel 8087 math coprocessor and is supported
+	by all processors that are based on the x86 design that incorporate
+	a floating-point unit(FPU).This 80 - bit format uses one bit for
+	the sign of the significand, 15 bits for the exponent field
+	(i.e. the same range as the 128 - bit quadruple precision IEEE 754 format)
+	and 64 bits for the significand. The exponent field is biased by 16383,
+	meaning that 16383 has to be subtracted from the value in the
+	exponent field to compute the actual power of 2.
+	An exponent field value of 32767 (all fifteen bits 1) is reserved
+	so as to enable the representation of special states such as
+	infinity and Not a Number.If the exponent field is zero, the
+	value is a denormal number and the exponent of 2 is ¿16382.
+*/
+#if defined(__clang__) || defined(__GNUC__) || defined(__GNUG__)
+/* Clang/LLVM. ---------------------------------------------- */
+/* GNU GCC/G++. --------------------------------------------- */
+
+#elif defined(__ICC) || defined(__INTEL_COMPILER)
+/* Intel ICC/ICPC. ------------------------------------------ */
+
+#elif defined(__HP_cc) || defined(__HP_aCC)
+/* Hewlett-Packard C/C++. ---------------------------------- */
+
+
+#elif defined(__IBMC__) || defined(__IBMCPP__)
+/* IBM XL C/C++. -------------------------------------------- */
+
+#elif defined(_MSC_VER)
+/* Microsoft Visual Studio. --------------------------------- */
+// Visual C++ compiler is 15.00.20706.01, the _MSC_FULL_VER will be 15002070601
+
+// Visual C++ does not support long double, it is just an alias for double
+
+#elif defined(__PGI)
+/* Portland Group PGCC/PGCPP. ------------------------------- */
+
+#elif defined(__SUNPRO_C) || defined(__SUNPRO_CC)
+/* Oracle Solaris Studio. ----------------------------------- */
+
+#elif defined(__riscv)
+/* RISC-V G++ tool chain */
+
+
+#endif
+
+} // namespace sw::universal
+
