+++ include/universal/native/ieee754_hpcc.hpp
@@ -1,0 +1,66 @@
+#pragma once
+// ieee754.hpp: HP C/C++ specific manipulation functions for IEEE-754 native types
+//
+// Copyright (C) 2017-2021 Stillwater Supercomputing, Inc.
+//
+// This file is part of the universal numbers project, which is released under an MIT Open Source license.
+#if defined(__HP_cc) || defined(__HP_aCC)
+/* Hewlett-Packard C/C++. ---------------------------------- */
+
+namespace sw::universal {
+
+////////////////////////////////////////////////////////////////////////
+// numerical helpers
+
+
+/////////////////////////////////////////////////////////////////////////////////////////////////////////
+// compiler specific long double IEEE floating point
+
+/*
+	Long double is not consistently implemented across different compilers.
+	The following section organizes the implementation details of each
+	of the compilers supported.
+
+	The x86 extended precision format is an 80-bit format first
+	implemented in the Intel 8087 math coprocessor and is supported
+	by all processors that are based on the x86 design that incorporate
+	a floating-point unit(FPU).This 80 - bit format uses one bit for
+	the sign of the significand, 15 bits for the exponent field
+	(i.e. the same range as the 128 - bit quadruple precision IEEE 754 format)
+	and 64 bits for the significand. The exponent field is biased by 16383,
+	meaning that 16383 has to be subtracted from the value in the
+	exponent field to compute the actual power of 2.
+	An exponent field value of 32767 (all fifteen bits 1) is reserved
+	so as to enable the representation of special states such as
+	infinity and Not a Number.If the exponent field is zero, the
+	value is a denormal number and the exponent of 2 is 16382.
+*/
+
+// generate a binary string for a native long double precision IEEE floating point
+inline std::string to_hex(const long double& number) {
+	return std::string("to_hex() not implemented for HP compiler");
+}
+
+// generate a binary string for a native long double precision IEEE floating point
+inline std::string to_binary(const long double& number, bool bNibbleMarker = false) {
+	return std::string("to_binary() not implemented for HP compiler");
+}
+
+// return in triple form (+, scale, fraction)
+inline std::string to_triple(const long double& number) {
+	return std::string("to_triple() not implemented for HP compiler");
+}
+
+inline void extract_fp_components(float fp, bool& _sign, int& _exponent, float& _fr, uint32_t& _fraction) {
+	std::cerr << "extract_fp_components not implemented for HP compiler");
+}
+inline void extract_fp_components(double fp, bool& _sign, int& _exponent, float& _fr, uint32_t& _fraction) {
+	std::cerr << "extract_fp_components not implemented for HP compiler");
+}
+inline void extract_fp_components(long double fp, bool& _sign, int& _exponent, float& _fr, uint32_t& _fraction) {
+	std::cerr << "extract_fp_components not implemented for HP compiler");
+}
+
+} // namespace sw::universal
+
+#endif // Hewlett-Packard C/C++.
